feat(classifier): subscription fallback, statusline visibility, rename to approval_classifier
Implements R5 of specs/2026-08-03-api-key-switching.md.

Two backends instead of one. The PermissionRequest classifier used to have a
single path: resolve ANTHROPIC_API_KEY, call the API, and on any failure fall
through to a manual permission prompt. A silently-empty key produced ~349
consecutive manual denials before anyone noticed. It now tries the API first
(~1s) and, on ANY failure (no key, 401/403, 429, provider 5xx, credits,
workspace limit), classifies via 'claude -p --model haiku' under the CLI's
OAuth login (~9s measured). Only if BOTH fail does the user get the prompt,
and then the warning names both failures so a broken login is not mistaken
for a broken key.

'claude -p --bare' is deliberately NOT the recursion guard: its help states
OAuth and keychain are never read, and it returns 'Not logged in' in practice,
so it is the one mode that cannot reach the subscription. Recursion is stopped
by the APPROVAL_CLASSIFIER_NESTED env marker, checked before stdin is read.
The nested call runs with cwd=~ so an untrusted repo's CLAUDE.md cannot inject
instructions into the prompt that decides whether to auto-approve actions in
that same repo, and with the failed key stripped from the environment.

Degradation is now visible instead of silent. The hook writes
~/.cache/claude/approval-classifier-health.json on every attempt (atomically --
the statusline may be mid-read), and the statusline renders auto:<key> when
healthy, auto:sub (<key> down) when degraded, and a red marker when dead.
Writing on successes too is what makes recovery automatic: restoring a key
returns the statusline to healthy on the next classification with no reset.
A health file older than 6h is treated as absent.

Renamed auto_classify.py to approval_classifier.py (and the rules file, the
log path, the exception class, and all user-facing strings). Sessions
repeatedly confused it with Anthropic's harness auto mode classifier, which is
an unrelated path; the new name also says what it does -- it auto-approves.

Both statusline implementations were updated, so three guards now cover their
drift: tests/test_statusline_classifier.sh asserts the Rust and bash renderers
produce byte-identical segments across all three backend states,
nudge_statusline_parity.sh fires on edits to either file, and
check-claude-tools-fresh.sh flags committed binaries older than the source.
That last one exists because darwin-arm64 cannot be cross-compiled from Linux,
so a stale Mac binary fails silently -- this commit ships a rebuilt
linux-x86_64 only; the Mac binary must be rebuilt there.

Verified with live calls, not mocks: broken key -> subscription allow in 9.7s;
healthy key -> api backend with automatic recovery; both broken -> loud
warning plus systemMessage. 15/15 statusline tests pass. New shell is
shellcheck-clean.
