+++ CLAUDE.md
@@ -67,3 +67,5 @@ Project-specific bugs, quirks, decisions and current state. Timestamp `- descrip
 
 - The simplify nudge now has a second signal: `simplify_track_reuse.py` counts *runs* of scratch scripts (not writes — writes are iteration) and the Stop hook suggests promoting any that ran 3× with at least one run unchanged. Rules for where a promoted script lands: `claude/rules/reusable-component-promotion.md` (2026-08-01)
 - Background-session job dirs (`~/.claude/jobs/<id>/`) have no built-in retention — `cleanupPeriodDays` only covers `~/.claude/projects` transcripts. `custom_bins/claude-jobs-reap` deletes terminal-state job dirs inactive >7d (active-looking states get 30d grace); `claude/hooks/reap_jobs.sh` runs it daily via SessionStart. Deleting a job dir removes the `claude agents` entry only — the transcript and `--resume` survive (2026-08-02)
+- **The PermissionRequest hook is `claude/hooks/approval_classifier.py`** (was `auto_classify.py`) — renamed because sessions kept confusing it with Anthropic's harness "auto mode classifier", which is a completely separate path. It now has two backends: the API key first (~1s), then the Claude subscription via `claude -p --model haiku` (~9s) when the API path fails for any reason. `claude -p --bare` cannot be used for the fallback — its help states OAuth and keychain are never read — so recursion is stopped by the `APPROVAL_CLASSIFIER_NESTED` env marker instead. State is written to `~/.cache/claude/approval-classifier-health.json` on every attempt and rendered by the statusline as `auto:<key>` / `auto:sub` / `🔴auto` (2026-08-03)
+- The statusline has **two implementations that must stay in sync**: `tools/claude-tools/src/statusline.rs` (what actually runs) and `claude/statusline.sh` (fallback). Guards added because comments alone can't catch drift: `tests/test_statusline_classifier.sh` asserts both render identically, `claude/hooks/nudge_statusline_parity.sh` fires on edits to either, and `scripts/check-claude-tools-fresh.sh` flags committed binaries older than the source — the silent failure is `darwin-arm64`, which nobody on Linux can cross-compile (2026-08-03)
